requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.17.2
    # via mlflow
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via starlette
blinker==1.9.0
    # via flask
cachetools==6.2.2
    # via
    #   google-auth
    #   mlflow-skinny
    #   mlflow-tracing
certifi==2025.11.12
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   mlflow-skinny
    #   uvicorn
cloudpickle==3.1.2
    # via mlflow-skinny
contourpy==1.3.3
    # via matplotlib
cryptography==46.0.3
    # via mlflow
cycler==0.12.1
    # via matplotlib
databricks-sdk==0.73.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
db-dtypes==1.4.4
    # via pandas-gbq
docker==7.1.0
    # via mlflow
duckdb==1.4.2
    # via mlflow-model-gcp (pyproject.toml)
fastapi==0.121.2
    # via mlflow-skinny
flask==3.1.2
    # via
    #   flask-cors
    #   mlflow
flask-cors==6.0.1
    # via mlflow
fonttools==4.60.1
    # via matplotlib
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via mlflow-skinny
google-api-core==2.28.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
    #   pandas-gbq
google-auth==2.43.0
    # via
    #   databricks-sdk
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-bigquery
    #   google-cloud-core
    #   pandas-gbq
    #   pydata-google-auth
google-auth-oauthlib==1.2.2
    # via
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery==3.38.0
    # via pandas-gbq
google-cloud-core==2.5.0
    # via google-cloud-bigquery
google-crc32c==1.7.1
    # via google-resumable-media
google-resumable-media==2.8.0
    # via google-cloud-bigquery
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
graphene==3.4.3
    # via mlflow
graphql-core==3.2.7
    # via
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
greenlet==3.2.4
    # via sqlalchemy
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
gunicorn==23.0.0
    # via mlflow
h11==0.16.0
    # via uvicorn
huey==2.5.4
    # via mlflow
idna==3.11
    # via
    #   anyio
    #   requests
importlib-metadata==8.7.0
    # via
    #   mlflow-skinny
    #   opentelemetry-api
iniconfig==2.3.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   pytest-html
joblib==1.5.2
    # via scikit-learn
kiwisolver==1.4.9
    # via matplotlib
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
matplotlib==3.10.7
    # via mlflow
mlflow==3.6.0
    # via mlflow-model-gcp (pyproject.toml)
mlflow-skinny==3.6.0
    # via mlflow
mlflow-tracing==3.6.0
    # via mlflow
numpy==2.3.5
    # via
    #   mlflow-model-gcp (pyproject.toml)
    #   contourpy
    #   db-dtypes
    #   matplotlib
    #   mlflow
    #   pandas
    #   pandas-gbq
    #   scikit-learn
    #   scipy
oauthlib==3.3.1
    # via requests-oauthlib
opentelemetry-api==1.38.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-proto==1.38.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
opentelemetry-sdk==1.38.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
opentelemetry-semantic-conventions==0.59b0
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   db-dtypes
    #   google-cloud-bigquery
    #   gunicorn
    #   matplotlib
    #   mlflow-skinny
    #   mlflow-tracing
    #   pandas-gbq
    #   pytest
pandas==2.3.3
    # via
    #   mlflow-model-gcp (pyproject.toml)
    #   db-dtypes
    #   mlflow
    #   pandas-gbq
pandas-gbq==0.31.0
    # via mlflow-model-gcp (pyproject.toml)
pillow==12.0.0
    # via matplotlib
pluggy==1.6.0
    # via pytest
proto-plus==1.26.1
    # via google-api-core
protobuf==6.33.1
    # via
    #   databricks-sdk
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   mlflow-skinny
    #   mlflow-tracing
    #   opentelemetry-proto
    #   proto-plus
psutil==7.1.3
    # via pandas-gbq
pyarrow==22.0.0
    # via
    #   db-dtypes
    #   mlflow
    #   pandas-gbq
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   fastapi
    #   mlflow-skinny
    #   mlflow-tracing
pydantic-core==2.41.5
    # via pydantic
pydata-google-auth==1.9.1
    # via pandas-gbq
pygments==2.19.2
    # via pytest
pyparsing==3.2.5
    # via matplotlib
pytest==9.0.1
    # via
    #   pytest-html
    #   pytest-metadata
pytest-html==4.1.1
    # via mlflow-model-gcp (pyproject.toml)
pytest-metadata==3.1.1
    # via pytest-html
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   graphene
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via mlflow-skinny
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via mlflow-skinny
requests==2.32.5
    # via
    #   databricks-sdk
    #   docker
    #   google-api-core
    #   google-cloud-bigquery
    #   mlflow-skinny
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
scikit-learn==1.7.2
    # via mlflow
scipy==1.16.3
    # via
    #   mlflow
    #   scikit-learn
setuptools==80.9.0
    # via
    #   pandas-gbq
    #   pydata-google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via
    #   alembic
    #   mlflow
sqlparse==0.5.3
    # via mlflow-skinny
starlette==0.49.3
    # via fastapi
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   graphene
    #   grpcio
    #   mlflow-skinny
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   docker
    #   requests
uvicorn==0.38.0
    # via mlflow-skinny
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
zipp==3.23.0
    # via importlib-metadata
